Move raw alasql querying and prepared version under test.
