+++ index.js
@@ -1,6 +1,8 @@
+const { filterAndTransform, prepareFilterAndTransform } = require('./lib/filter-and-transform')
+
 module.exports = {
   query: require('./lib/query'),
   prepareQuery: require('./lib/prepare-query'),
-  querySql: require('./lib/query-sql'),
-  prepareSql: require('./lib/prepare-sql')
+  querySql: filterAndTransform,
+  prepareSql: prepareFilterAndTransform
 }
