+++ lib/execute-query.js
@@ -1,5 +1,5 @@
 'use strict'
-const sql = require('./filter-and-transform')
+const { filterAndTransform: sql } = require('./filter-and-transform')
 const Query = require('./sql-query-builder')
 const { calculateClassBreaks, calculateUniqueValueBreaks } = require('./generateBreaks/index')
