--- lib/query-sql.js
@@ -1,5 +1,0 @@
-const filterAndTransform = require('./filter-and-transform')
-
-module.exports = function (statement, params) {
-  return filterAndTransform(statement, params)
-}
